Styling sign up finish, share form components
