+++ src/screens/SignupScreen.js
@@ -1,23 +1,21 @@
 import React, {Component} from 'react';
-import {Redirect, withRouter} from "react-router";
-import {FormattedMessage, injectIntl} from "react-intl";
-import {Button, Intent} from '@blueprintjs/core';
-import {endpoint} from "../api";
-import {xhrErrorToast, xhrSuccessToast} from "../components/XhrToast";
-import {connect} from "react-redux";
-import {login} from "../actions/sessionActions";
-import mapValues from 'lodash/mapValues';
+import {connect} from 'react-redux';
+import {Redirect} from 'react-router';
+import {injectIntl} from 'react-intl';
+
+import {endpoint} from '../api';
+import {login} from '../actions/sessionActions';
+
+import PasswordSignup from '../components/PasswordSignup';
+import {xhrErrorToast, xhrSuccessToast} from '../components/XhrToast';
 
 class SignupScreen extends Component {
   elements = {};
 
-  submit(event) {
+  submit(data) {
     const {match: {params}, intl, metadata, login} = this.props;
-    event.preventDefault();
 
-    const data = {code: params.code, ...mapValues(this.elements, 'value')};
-
-    endpoint.post('/roles', data).then((res) => {
+    endpoint.post('/roles', {code: params.code, ...data}).then((res) => {
       xhrSuccessToast(res, intl);
       endpoint.post(metadata.auth.password_login_uri, data).then((res) => {
         login(res.data.token);
@@ -41,26 +39,12 @@ class SignupScreen extends Component {
       return <Redirect to="/"/>;
     }
 
-    return <section>
-      <form onSubmit={(e) => this.submit(e)}>
-        <label className="pt-label">
-          <FormattedMessage id="signup.email" defaultMessage="E-Mail address"/>
-          <input className="pt-input" type="email" name="email" required ref={(el) => this.elements["email"] = el}/>
-        </label>
-        <label className="pt-label">
-          <FormattedMessage id="signup.name" defaultMessage="Your Name"/>
-          <input className="pt-input" type="text" name="name" required ref={(el) => this.elements["name"] = el}/>
-        </label>
-        <label className="pt-label">
-          <FormattedMessage id="signup.password" defaultMessage="Password"/>
-          <input className="pt-input" type="password" name="password" required
-                 ref={(el) => this.elements["password"] = el}/>
-        </label>
-        <Button iconName="log-in" intent={Intent.PRIMARY} type="submit">
-          <FormattedMessage id="signup.submit" defaultMessage="Signup"/>
-        </Button>
-      </form>
-    </section>
+    return (
+      <section className="small-screen">
+        <h2>Complete sign up</h2>
+        <PasswordSignup onSignup={data => this.submit(data)} showFull/>
+      </section>
+    );
   }
 }
 
@@ -69,6 +53,6 @@ const mapStateToProps = ({session, metadata}) => ({session, metadata});
 SignupScreen = connect(
   mapStateToProps,
   {login}
-)(withRouter(injectIntl(SignupScreen)));
+)(injectIntl(SignupScreen));
 
 export default SignupScreen;
